+++ cf/mixin/properties.py
@@ -660,7 +660,7 @@ class Properties(Container):
             The name of the property.
 
             *Parameter example:*
-               ``prop='long_name'``
+              ``prop='long_name'``
 
         default: optional
             Return the value of the *default* parameter if the