+++ cf/mixin/propertiesdata.py
@@ -2728,114 +2728,6 @@ class PropertiesData(Properties):
 
         return data.count_masked()
 
-    def creation_commands(self, representative_data=False,
-                          namespace='cf', indent=0, string=True,
-                          name='c', data_name='d'):
-        '''Return the commands that would create the construct.
-
-    .. versionadded:: 3.2.0
-
-    .. seealso:: `cf.Data.creation_commands`,
-                 `cf.Field.creation_commands`
-
-    :Parameters:
-
-        representative_data: `bool`, optional
-            Return one-line representations of `Data` instances, which
-            are not executable code but prevent the data being
-            converted in its entirety to a string representation.
-
-        namespace: `str`, optional
-            The namespace containing classes of the ``cf``
-            package. This is prefixed to the class name in commands
-            that instantiate instances of ``cf`` objects. By default,
-            *namespace* is ``'cf'``, i.e. it is assumed that ``cf``
-            was imported as ``import cf``.
-
-            *Parameter example:*
-              If ``cf`` was imported as ``import cf as cfp`` then set
-              ``namespace='cfp'``
-
-            *Parameter example:*
-              If ``cf`` was imported as ``from cf import *`` then set
-              ``namespace=''``
-
-        indent: `int`, optional
-            Indent each line by this many spaces. By default no
-            indentation is applied. Ignored if *string* is False.
-
-        string: `bool`, optional
-            If False then return each command as an element of a
-            `list`. By default the commands are concatenated into
-            a string, with a new line inserted between each command.
-
-    :Returns:
-
-        `str` or `list`
-            The commands in a string, with a new line inserted between
-            each command. If *string* is False then the separate
-            commands are returned as each element of a `list`.
-
-    **Examples:**
-
-        TODO
-
-        '''
-        if name == data_name:
-            raise ValueError(
-                "'name' and 'data_name' parameters can not have the same "
-                "value: {!r}".format(name)
-            )
-
-        namespace0 = namespace
-        if namespace0:
-            namespace = namespace+"."
-        else:
-            namespace = ""
-
-        indent = ' ' * indent
-
-        out = []
-
-        construct_type = getattr(self, 'construct_type', None)
-        if construct_type is not None:
-            out.append("# {}: {}".format(construct_type,
-                                         self.identity()))
-
-        out.append("{} = {}{}()".format(name, namespace,
-                                        self.__class__.__name__))
-
-        properties = self.properties()
-        if properties:
-            for prop in self.inherited_properties():
-                properties.pop(prop, None)
-
-            out.append("{}.set_properties({})".format(name,
-                                                      properties))
-
-        data = self.get_data(None)
-        if data is not None:
-            if representative_data:
-                out.append("{} = {!r} # Representative data".format(
-                    data_name, data))
-            else:
-                out.extend(data.creation_commands(name=data_name,
-                                                  namespace=namespace0,
-                                                  indent=0,
-                                                  string=False))
-
-            out.append("{}.set_data(d)".format(name))
-
-        nc = self.nc_get_variable(None)
-        if nc is not None:
-            out.append("{}.nc_set_variable({!r})".format(name, nc))
-
-        if string:
-            out[0] = indent+out[0]
-            out = ('\n'+indent).join(out)
-
-        return out
-
     def cyclic(self, axes=None, iscyclic=True):
         '''Set the cyclicity of an axis.
 
@@ -4629,11 +4521,11 @@ class PropertiesData(Properties):
     By default the identity is the first found of the following:
 
     * The "standard_name" property.
-    * The "id" attribute, preceeded by ``'id%'``.
-    * The "cf_role" property, preceeded by ``'cf_role='``.
-    * The "axis" property, preceeded by ``'axis='``.
-    * The "long_name" property, preceeded by ``'long_name='``.
-    * The netCDF variable name, preceeded by ``'ncvar%'``.
+    * The "id" attribute, preceded by ``'id%'``.
+    * The "cf_role" property, preceded by ``'cf_role='``.
+    * The "axis" property, preceded by ``'axis='``.
+    * The "long_name" property, preceded by ``'long_name='``.
+    * The netCDF variable name, preceded by ``'ncvar%'``.
     * The coordinate type (``'X'``, ``'Y'``, ``'Z'`` or ``'T'``).
     * The value of the *default* parameter.
 
@@ -4756,14 +4648,14 @@ class PropertiesData(Properties):
     The identities comprise:
 
     * The "standard_name" property.
-    * The "id" attribute, preceeded by ``'id%'``.
-    * The "cf_role" property, preceeded by ``'cf_role='``.
-    * The "axis" property, preceeded by ``'axis='``.
-    * The "long_name" property, preceeded by ``'long_name='``.
-    * All other properties (including "standard_name"), preceeded by
+    * The "id" attribute, preceded by ``'id%'``.
+    * The "cf_role" property, preceded by ``'cf_role='``.
+    * The "axis" property, preceded by ``'axis='``.
+    * The "long_name" property, preceded by ``'long_name='``.
+    * All other properties (including "standard_name"), preceded by
       the property name and an ``'='``.
     * The coordinate type (``'X'``, ``'Y'``, ``'Z'`` or ``'T'``).
-    * The netCDF variable name, preceeded by ``'ncvar%'``.
+    * The netCDF variable name, preceded by ``'ncvar%'``.
 
     .. versionadded:: 3.0.0